Add files via upload
When using this, the database should probably be named Dragonfruit. If you name it something different then change the first line in the SQL query from Use Dragonfruit to Use {insert DatabaseName} (without the brackets). When running and making the tables in the database, run Options, BeverageInfo, SidesInfo, Customer, and Manager first. Then run BeveragesInfo and SidesInfo, Pizza, CustomPizzas, Employee, Orders, and Payments in that order. To view the tables, use Select * from {insert table name} (without the brackets) and highlight only that phrase or open another query and paste that query in. Let me know if you have any more questions.