Add files via upload
This and the HTML file from 11.17.24 are the newest ones. It needs to be connected properly and tested more. I figured out the main logic though so that should be good. I used the size of the pizza to determine the cost and if they have more than one topping it adds on depending on if it's meat or vegetables (there are 5 of each). You will have to finish up the addtoCart function. I put the basic idea in there. Just one pizza is being added at a time and each one has a unique ID associated with it that is then pushed to the orders table as part of larger order for one customer with a unique ID for each order and customer. Customer and OrderID are used as the primary composite key for the order table. For the styling, I couldn't connect my CSS file so the most recent one in this pages branch is fine, I just used in-line CSS. Also take a look at the form action for the pizza HTML file, I couldn't figure that out. Let me know if you have any questions.